// dashboard/src/api/models/taxId.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * Invoicence API
 * Description
 * OpenAPI spec version: 1.0.0
 */
import type { TaxIdCountry } from "./taxIdCountry";
import type { TaxIdTypeEnum } from "./taxIdTypeEnum";

export interface TaxId {
  id: string;
  type: TaxIdTypeEnum;
  number: string;
  /** @nullable */
  country: TaxIdCountry;
  created_at: string;
  updated_at: string;
}